program-runtime: add `CallerAccount` api from syscall lib
